+++ exporters/autoexport/spans.go
@@ -13,7 +13,12 @@ import (
 	"go.opentelemetry.io/otel/sdk/trace"
 )
 
-const otelExporterOTLPTracesProtoEnvKey = "OTEL_EXPORTER_OTLP_TRACES_PROTOCOL"
+const (
+	otelTracesExporterEnvKey         = "OTEL_TRACES_EXPORTER"
+	otelTracesExporterProtocolEnvKey = "OTEL_EXPORTER_OTLP_TRACES_PROTOCOL"
+)
+
+var tracesSignal = newSignal[trace.SpanExporter](otelTracesExporterEnvKey)
 
 // SpanOption applies an autoexport configuration option.
 type SpanOption = option[trace.SpanExporter]
@@ -23,14 +28,36 @@ type SpanOption = option[trace.SpanExporter]
 // Deprecated: Use SpanOption.
 type Option = SpanOption
 
-// WithFallbackSpanExporter sets the fallback exporter to use when no exporter
-// is configured through the OTEL_TRACES_EXPORTER environment variable.
-func WithFallbackSpanExporter(spanExporterFactory func(ctx context.Context) (trace.SpanExporter, error)) SpanOption {
-	return withFallbackFactory[trace.SpanExporter](spanExporterFactory)
+// NewSpanExporters returns one or more configured [go.opentelemetry.io/otel/sdk/trace.SpanExporter]
+// defined using the environment variables described below.
+//
+// OTEL_TRACES_EXPORTER defines the traces exporter; this value accepts a comma-separated list of values; supported values:
+//   - "none" - "no operation" exporter
+//   - "otlp" (default) - OTLP exporter; see [go.opentelemetry.io/otel/exporters/otlp/otlptrace]
+//   - "console" - Standard output exporter; see [go.opentelemetry.io/otel/exporters/stdout/stdouttrace]
+//
+// OTEL_EXPORTER_OTLP_PROTOCOL defines OTLP exporter's transport protocol;
+// supported values:
+//   - "grpc" - protobuf-encoded data using gRPC wire format over HTTP/2 connection;
+//     see: [go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc]
+//   - "http/protobuf" (default) -  protobuf-encoded data over HTTP connection;
+//     see: [go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp]
+//
+// An error is returned if an environment value is set to an unhandled value.
+// Use [WithFallbackSpanExporter] option to change the returned exporter
+// when OTEL_TRACES_EXPORTER is unset or empty.
+//
+// Use [RegisterSpanExporter] to handle more values of OTEL_TRACES_EXPORTER.
+//
+// Use [IsNoneSpanExporter] to check if the returned exporter is a "no operation" exporter.
+func NewSpanExporters(ctx context.Context, options ...SpanOption) ([]trace.SpanExporter, error) {
+	return tracesSignal.create(ctx, options...)
 }
 
 // NewSpanExporter returns a configured [go.opentelemetry.io/otel/sdk/trace.SpanExporter]
 // defined using the environment variables described below.
+//
+// DEPRECATED: consider using [NewSpanExporters] instead.
 //
 // OTEL_TRACES_EXPORTER defines the traces exporter; supported values:
 //   - "none" - "no operation" exporter
@@ -48,15 +75,18 @@ func WithFallbackSpanExporter(spanExporterFactory func(ctx context.Context) (tra
 // supported values are the same as OTEL_EXPORTER_OTLP_PROTOCOL.
 //
 // An error is returned if an environment value is set to an unhandled value.
-//
-// Use [RegisterSpanExporter] to handle more values of OTEL_TRACES_EXPORTER.
-//
 // Use [WithFallbackSpanExporter] option to change the returned exporter
 // when OTEL_TRACES_EXPORTER is unset or empty.
 //
+// Use [RegisterSpanExporter] to handle more values of OTEL_TRACES_EXPORTER.
+//
 // Use [IsNoneSpanExporter] to check if the returned exporter is a "no operation" exporter.
 func NewSpanExporter(ctx context.Context, opts ...SpanOption) (trace.SpanExporter, error) {
-	return tracesSignal.create(ctx, opts...)
+	exporters, err := NewSpanExporters(ctx, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return exporters[0], nil
 }
 
 // RegisterSpanExporter sets the SpanExporter factory to be used when the
@@ -66,11 +96,15 @@ func RegisterSpanExporter(name string, factory func(context.Context) (trace.Span
 	must(tracesSignal.registry.store(name, factory))
 }
 
-var tracesSignal = newSignal[trace.SpanExporter]("OTEL_TRACES_EXPORTER")
+// WithFallbackSpanExporter sets the fallback exporter to use when no exporter
+// is configured through the OTEL_TRACES_EXPORTER environment variable.
+func WithFallbackSpanExporter(spanExporterFactory func(ctx context.Context) (trace.SpanExporter, error)) SpanOption {
+	return withFallbackFactory[trace.SpanExporter](spanExporterFactory)
+}
 
 func init() {
 	RegisterSpanExporter("otlp", func(ctx context.Context) (trace.SpanExporter, error) {
-		proto := os.Getenv(otelExporterOTLPTracesProtoEnvKey)
+		proto := os.Getenv(otelTracesExporterProtocolEnvKey)
 		if proto == "" {
 			proto = os.Getenv(otelExporterOTLPProtoEnvKey)
 		}
@@ -89,10 +123,10 @@ func init() {
 			return nil, errInvalidOTLPProtocol
 		}
 	})
-	RegisterSpanExporter("console", func(ctx context.Context) (trace.SpanExporter, error) {
+	RegisterSpanExporter("console", func(_ context.Context) (trace.SpanExporter, error) {
 		return stdouttrace.New()
 	})
-	RegisterSpanExporter("none", func(ctx context.Context) (trace.SpanExporter, error) {
+	RegisterSpanExporter("none", func(_ context.Context) (trace.SpanExporter, error) {
 		return noopSpanExporter{}, nil
 	})
 }